[Refactor] Moved helper functions into db index.  Created test folder
