+++ package.json
@@ -9,14 +9,17 @@
   },
   "scripts": {
     "start": "nodemon server/index.js",
-    "seed-database": "node ./dummydata/seed.js"
+    "seed-database": "node ./dummydata/seed.js",
+    "test": "jest --watchAll --verbose"
   },
   "dependencies": {
+    "axios": "^0.18.0",
     "body-parser": "^1.18.3",
     "debug": "^4.1.0",
     "dotenv": "^6.1.0",
     "express": "^4.16.4",
     "faker": "^4.1.0",
+    "jest": "^23.6.0",
     "morgan": "^1.9.1",
     "pg": "^7.6.1",
     "pg-promise": "^8.5.2"
