+++ router/index.js
@@ -1,8 +1,9 @@
 const express = require('express');
 const router = express.Router();
-const pg = require('../helpers');
+const db = require('../db');
 
 router.get('/', async(req, res) => {
+  console.log('inside get!');
   const reviews = await pg.getReviews();
   res.json(reviews);
 });
